Restrict user login if the user type is "Other" Rename MessageConstants.kt file to ResourceConstants.kt
