+++ user-android-app/app/src/main/java/lnbti/charithgtp01/smartattendanceuserapp/constants/ResourceConstants.kt
@@ -1,10 +1,12 @@
 package lnbti.charithgtp01.smartattendanceuserapp.constants
 
-object MessageConstants {
+object ResourceConstants {
     const val NO_INTERNET = "No Internet"
     const val USER_REGISTERED_SUCCESS = "User Registered Successfully"
     const val OTHER_SETTINGS = "Other Settings"
     const val GET_DEVICE_ID = "Get Device ID"
     const val CHANGE_PASSWORD = "Change Password"
+    const val ANDROID_USER = "Android User"
+    const val OTHER = "Other"
 
 }
