+++ user-android-app/app/src/main/java/lnbti/charithgtp01/smartattendanceuserapp/ui/register/RegisterViewModel.kt
@@ -9,19 +9,12 @@ import com.google.android.material.textfield.TextInputEditText
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import lnbti.charithgtp01.smartattendanceuserapp.R
-import lnbti.charithgtp01.smartattendanceuserapp.constants.Constants
-import lnbti.charithgtp01.smartattendanceuserapp.constants.MessageConstants
-import lnbti.charithgtp01.smartattendanceuserapp.constants.MessageConstants.NO_INTERNET
-import lnbti.charithgtp01.smartattendanceuserapp.constants.MessageConstants.USER_REGISTERED_SUCCESS
-import lnbti.charithgtp01.smartattendanceuserapp.interfaces.CustomAlertDialogListener
+import lnbti.charithgtp01.smartattendanceuserapp.constants.ResourceConstants.NO_INTERNET
 import lnbti.charithgtp01.smartattendanceuserapp.model.ApiCallResponse
 import lnbti.charithgtp01.smartattendanceuserapp.model.Company
 import lnbti.charithgtp01.smartattendanceuserapp.model.RegisterRequest
 import lnbti.charithgtp01.smartattendanceuserapp.repositories.UserRepository
-import lnbti.charithgtp01.smartattendanceuserapp.ui.login.LoginActivity
-import lnbti.charithgtp01.smartattendanceuserapp.utils.DialogUtils
 import lnbti.charithgtp01.smartattendanceuserapp.utils.NetworkUtils
-import lnbti.charithgtp01.smartattendanceuserapp.utils.Utils
 import lnbti.charithgtp01.smartattendanceuserapp.utils.Validations.Companion.isEmailValid
 import lnbti.charithgtp01.smartattendanceuserapp.utils.Validations.Companion.isMobileNumberValid
 import lnbti.charithgtp01.smartattendanceuserapp.utils.Validations.Companion.isNICValid
